requirements.txt
distlib==0.3.4
filelock==3.6.0
importlib-metadata==4.11.3
lxml==4.8.0
Markdown==3.3.6
markdown2==2.4.2
platformdirs==2.5.1
python-docx==0.8.11
virtualenv==20.14.0
zipp==3.8.0
